raspi-gpio/requirements.txt
###### Requirements without Version Specifiers ######

paho-mqtt
gpiozero
mysql.connector

###### Requirements with Version Specifiers ######
